refactor: 테스트 픽스쳐 한 곳으로 이동 및 given when then 키워드 삽입
